1.修复了上电不启动的BUG 2.将数值由小数显示，单位N*m

+++ User/demo.h
@@ -8,7 +8,7 @@ typedef union
 	unsigned char c[4];		
 }Data_Processing_U;
 
-extern float Num;
+extern int Num;
 
 void SendRequst(void);
 
@@ -25,7 +25,7 @@ void RGY_Light_Task(void);
 
 void Speaker_Task(void);
 
-extern float show_num;	
+extern int show_num;	
 extern uint8_t ReadMeasuredValue[];
 extern Data_Processing_U Torque_value;//±äÁ¿ÉùÃ÷
 #endif
